style: left-align person bio
